Add files via upload
So sometimes this works (Well at the end when it asks for user input it will break) but other times the server will send messages too fast and the client receives two messages in 1 (specifically 1file1 is a possible error). But this still kinda has the flow I was talking about when sending and receiving the contents of the file.